# openapi.yaml
openapi: 3.1.0
info:
  title: auth_service API
  version: 0.1.0
  description: |
    Контрактный API для identity и сессий платформы.
    Поддерживает login/refresh/logout/me и публикацию JWKS.
servers:
  - url: http://localhost:8000

tags:
  - name: health
  - name: auth
  - name: public

paths:
  /healthz:
    get:
      tags: [health]
      summary: Проверка состояния
      operationId: healthCheck
      responses:
        '200':
          description: Сервис доступен
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok

  /v1/auth/login:
    post:
      tags: [auth]
      summary: Вход пользователя
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPairResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }

  /v1/auth/register:
    post:
      tags: [auth]
      summary: Регистрация пользователя
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Учетная запись создана
          content:
            application/json:
              schema:
                type: object
                required: [account_id, user_id]
                properties:
                  account_id:
                    type: string
                    format: uuid
                  user_id:
                    type: string
                    format: uuid
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }

  /v1/auth/refresh:
    post:
      tags: [auth]
      summary: Обновление access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Токены обновлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPairResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }

  /v1/auth/logout:
    post:
      tags: [auth]
      summary: Завершение сессии
      operationId: logout
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '204':
          description: Сессия завершена
        '401': { $ref: '#/components/responses/Unauthorized' }

  /v1/auth/me:
    get:
      tags: [auth]
      summary: Профиль текущего пользователя
      operationId: me
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }

  /v1/auth/sessions:
    get:
      tags: [auth]
      summary: Список сессий текущего пользователя
      operationId: listMySessions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Список сессий
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionListResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }

  /.well-known/jwks.json:
    get:
      tags: [public]
      summary: Публичные ключи JWKS
      operationId: getJwks
      responses:
        '200':
          description: Документ JWKS
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwksResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    Unauthorized:
      description: Не авторизован
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Конфликт или нарушение инварианта
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ValidationError:
      description: Ошибка валидации
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'

  schemas:
    Problem:
      type: object
      required: [type, title, status, detail]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        request_id: { type: string }

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        ip_address:
          type: string
          nullable: true
        user_agent_raw:
          type: string
          nullable: true
        device_type:
          type: string
          enum: [desktop, mobile, tablet, bot, unknown]
          nullable: true
        os_name:
          type: string
          nullable: true
        os_version:
          type: string
          nullable: true
        browser_name:
          type: string
          nullable: true
        browser_version:
          type: string
          nullable: true
        client_name:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        auth_method:
          type: string
          enum: [password, refresh, oauth, magic_link]
        mfa_used:
          type: boolean
        is_trusted:
          type: boolean
        risk_level:
          type: string
          enum: [low, medium, high]
        session_fingerprint:
          type: string
          nullable: true
        last_path:
          type: string
          nullable: true
        last_action:
          type: string
          nullable: true

    RegisterRequest:
      type: object
      required: [email, password, default_role]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        default_role:
          type: string
          enum: [admin, teacher, parent, student]

    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string

    LogoutRequest:
      type: object
      required: [session_id]
      properties:
        session_id:
          type: string
          format: uuid

    TokenPairResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          const: Bearer
        expires_in:
          type: integer
          minimum: 1

    MeResponse:
      type: object
      required: [account_id, user_id, email, roles, status]
      properties:
        account_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        roles:
          type: array
          items:
            type: string
            enum: [admin, teacher, parent, student]
        status:
          type: string
          enum: [active, blocked, archived]

    JwksResponse:
      type: object
      required: [keys]
      properties:
        keys:
          type: array
          items:
            type: object
            additionalProperties: true

    SessionItemResponse:
      type: object
      required:
        [session_id, user_id, account_id, status, auth_method, mfa_used, is_trusted, risk_level, session_fingerprint, request_count, created_at]
      properties:
        session_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [active, closed, revoked]
        auth_method:
          type: string
          enum: [password, refresh, oauth, magic_link]
        mfa_used:
          type: boolean
        is_trusted:
          type: boolean
        risk_level:
          type: string
          enum: [low, medium, high]
        session_fingerprint:
          type: string
        request_count:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
        last_seen_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        revoked_at:
          type: string
          format: date-time
          nullable: true
        revoke_reason:
          type: string
          nullable: true
        ip_address:
          type: string
          nullable: true
        user_agent_raw:
          type: string
          nullable: true
        device_type:
          type: string
          enum: [desktop, mobile, tablet, bot, unknown]
          nullable: true
        os_name:
          type: string
          nullable: true
        os_version:
          type: string
          nullable: true
        browser_name:
          type: string
          nullable: true
        browser_version:
          type: string
          nullable: true
        client_name:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        last_path:
          type: string
          nullable: true
        last_action:
          type: string
          nullable: true

    SessionListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SessionItemResponse'